Copy compose from ca1394b6899

+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/platform/DesktopTextInputService2.kt
@@ -16,6 +16,7 @@
 
 package androidx.compose.ui.platform
 
+import androidx.compose.runtime.key
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.awt.toAwtRectangleRounded
 import androidx.compose.ui.scene.ComposeSceneMediator
@@ -154,7 +155,7 @@ private class InputMethodSession(
     // For our editor component we need this workaround.
     // After https://bugs.openjdk.java.net/browse/JDK-8074882 is fixed, this workaround should be replaced with a proper solution.
     var charKeyPressed: Boolean = false
-    private var needToDeletePreviousChar: Boolean = false
+    var needToDeletePreviousChar: Boolean = false
 
     override fun getLocationOffset(x: Int, y: Int): TextHitInfo? {
         if (composition != null) {
@@ -236,10 +237,10 @@ private class InputMethodSession(
     }
 
     fun inputMethodTextChanged(event: InputMethodEvent) {
-        val committedText = event.committedText
-        val composingText = event.composingText
+        val committed = event.committedText
+        val composing = event.composingText
 
-        imeComposingText = composingText
+        imeComposingText = composing
 
         if (ignoreNextInputMethodEvent) {
             ignoreNextInputMethodEvent = false
@@ -247,15 +248,13 @@ private class InputMethodSession(
         }
 
         request.editText {
-            if (needToDeletePreviousChar) {
-                if ((selection.min > 0) && composingText.isEmpty() && (composition == null)) {
-                    deleteSurroundingTextInCodePoints(1, 0)
-                }
+            if (needToDeletePreviousChar && selection.min > 0 && composing.isEmpty()) {
                 needToDeletePreviousChar = false
+                deleteSurroundingTextInCodePoints(1, 0)
             }
-            commitText(committedText, 1)
-            if (composingText.isNotEmpty()) {
-                setComposingText(composingText, 1)
+            commitText(committed, 1)
+            if (composing.isNotEmpty()) {
+                setComposingText(composing, 1)
             }
         }
     }
